requirements.txt
watchdog
python-dotenv
zat
pyod
